requirements.txt
# SO8T Safe Agent Requirements
# Core dependencies for training and inference

# PyTorch and CUDA support
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
--index-url https://download.pytorch.org/whl/cu121

# Transformers and model libraries
transformers>=4.35.0
tokenizers>=0.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
pyarrow>=12.0.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Bayesian optimization
optuna>=3.4.0
optuna-dashboard>=0.13.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.8.0
psutil>=5.9.0
GPUtil>=1.4.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# System utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0
typer>=0.9.0

# Audio notification (Windows: winsound is built-in, Linux: aplay/afplay)
# Windows: winsound (built-in)
# Linux: aplay (alsa-utils) or afplay (macOS)

# Optional: GGUF support (requires llama.cpp)
# llama-cpp-python>=0.2.0

# Optional: Additional model support
# openai>=1.0.0
# anthropic>=0.7.0

# Optional: Database support
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# Optional: Message queues
# redis>=4.6.0
# celery>=5.3.0

# Optional: Monitoring and observability
# prometheus-client>=0.17.0
# grafana-api>=1.0.3

# Optional: Security
# cryptography>=41.0.0
# bcrypt>=4.0.0

# Optional: Cloud storage
# boto3>=1.28.0
# azure-storage-blob>=12.17.0
# google-cloud-storage>=2.10.0